+++ test/test_CAN/test_can_protocol.cpp
@@ -3,8 +3,9 @@
 #include <unity.h>
 #include <ArduinoFake.h>
 #include <ArduinoJson.h>
-
 #include <canBus.hpp>
+
+#include "../mockSetup.hpp"
 
 using namespace fakeit;
 using namespace std;
@@ -13,9 +14,33 @@ using namespace std;
 
 #define CANFile "build/CAN/English/commands_hpsu.json"
 
+CAN_Config* TestCANConfig = nullptr;
+
 void setUp(void)
 {
     ArduinoFakeReset();
+
+    TestCANConfig = new CAN_Config();
+
+    ifstream inputFile(CANFile);
+
+    if(!inputFile) {
+        throw runtime_error("Run script \"build_can_commands.py\" before running test!");
+    }
+
+    string jsonContent((istreambuf_iterator<char>(inputFile)), istreambuf_iterator<char>());
+    inputFile.close();
+
+    DynamicJsonDocument configDoc(MODELS_CONFIG_SIZE);
+    deserializeJson(configDoc, jsonContent);
+
+    JsonObject jsonObject = configDoc.as<JsonObject>();
+
+    jsonObject["COMMANDS"] = jsonObject["Commands"];
+    jsonObject.remove("Commands");
+
+    canBus_fill_config(jsonObject, TestCANConfig);
+    //canBus_setup(TestCANConfig);
 }
 
 void tearDown(void)
